requirements.txt
Django==4.1.5
psycopg2==2.9.6
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
sendgrid==6.9.7
celery==5.2.7
redis==4.4.2
django-cors-headers==3.14.0
pandas==2.0.2
djangorestframework-api-key==2.3.0
